test: Move UIImage extension to its own file
